refactor(gateway): extract some functions
